Correções em units para compilação em versões do delphi sobre o  containstr e correção da nfs-e que mudou o tipo quantidade para currency (acho que deveria ser double) e ficou me outro local integer
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk@8349 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrNFSe/ACBrNFSeDANFSeFRDM.pas
@@ -192,7 +192,7 @@ begin
       begin
         Append;
         cdsItensServicoDiscriminacaoServico.AsString := Descricao;
-        cdsItensServicoQuantidade.AsInteger          := Quantidade;
+        cdsItensServicoQuantidade.AsFloat            := Quantidade;
         cdsItensServicoValorUnitario.AsFloat         := ValorUnitario;
         cdsItensServicoValorTotal.AsFloat            := ValorTotal;
         cdsItensServicoTributavel.AsString           := SimNaoToStr(Tributavel);
